+++ pages/expenceCat.py
@@ -1,0 +1,3 @@
+import streamlit as st
+
+st.subheader("Expense Categorization - Test")
